Feat#6 dongan

+++ jumun/src/main/java/com/mihak/jumun/entity/Category.java
@@ -1,6 +1,7 @@
 package com.mihak.jumun.entity;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 import javax.persistence.*;
 
@@ -12,8 +13,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CATEGORY_ID")
-    private Integer id;
+    private Long id;
 
     private String name;
-
 }
